HHH-14808 Fix PredicateSmokeTest
Refactors SpatialTestBase to use a SessionFactoryAware
implementation that performs necessary initialisations
(e.g. H2).
