+++ hibernate-spatial/src/test/java/org/hibernate/spatial/testing/SpatialTestBase.java
@@ -7,13 +7,7 @@
 
 package org.hibernate.spatial.testing;
 
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.util.Set;
-
 import org.hibernate.dialect.H2Dialect;
-import org.hibernate.spatial.CommonSpatialFunction;
-import org.hibernate.spatial.integration.SpatialTestDataProvider;
 import org.hibernate.spatial.testing.datareader.TestSupport;
 import org.hibernate.spatial.testing.domain.GeomEntity;
 import org.hibernate.spatial.testing.domain.JtsGeomEntity;
@@ -22,36 +16,14 @@ import org.hibernate.spatial.testing.domain.SpatialDomainModel;
 import org.hibernate.testing.orm.junit.DialectContext;
 import org.hibernate.testing.orm.junit.DomainModel;
 import org.hibernate.testing.orm.junit.SessionFactoryScope;
-import org.hibernate.testing.orm.junit.SessionFactoryScopeAware;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 
-import org.h2gis.functions.factory.H2GISFunctions;
-
 @DomainModel(modelDescriptorClasses = SpatialDomainModel.class)
 abstract public class SpatialTestBase
-		extends SpatialTestDataProvider implements SessionFactoryScopeAware {
-
-	protected SessionFactoryScope scope;
-	protected Set<String> supportedFunctions;
+		extends SpatialSessionFactoryAware  {
 
 	public abstract TestSupport.TestDataPurpose purpose();
-
-	@Override
-	public void injectSessionFactoryScope(SessionFactoryScope scope) {
-		this.scope = scope;
-		//scope is set to null during test cleanup
-		if ( scope != null ) {
-			this.supportedFunctions = scope.getSessionFactory()
-					.getQueryEngine()
-					.getSqmFunctionRegistry()
-					.getFunctions()
-					.keySet();
-			if ( DialectContext.getDialect() instanceof H2Dialect ) {
-				initH2GISExtensionsForInMemDb();
-			}
-		}
-	}
 
 	@BeforeEach
 	public void beforeEach() {
@@ -72,21 +44,5 @@ abstract public class SpatialTestBase
 		scope.inTransaction( session -> session.createQuery( "delete from JtsGeomEntity" ).executeUpdate() );
 	}
 
-	public boolean isSupported(CommonSpatialFunction function) {
-		return supportedFunctions.contains( function.name() );
-	}
-
-
-	private void initH2GISExtensionsForInMemDb() {
-		this.scope.inSession( session -> {
-			try {
-				Connection cn = session.getJdbcConnectionAccess().obtainConnection();
-				H2GISFunctions.load( cn );
-			}
-			catch (SQLException e) {
-				throw new RuntimeException( e );
-			}
-		} );
-	}
 
 }
